Create Skin Cancer Classification using Deep Learning
Skin cancer, especially melanoma is one of most deadly diseases. In the color images of skin, there is a high similarity between different skin lesion like melanoma and nevus, which increase the difficulty of the detection and diagnosis. A reliable automated system for skin lesion classification is essential for early detection to save effort, time and human life. In this project, an automated skin lesion classification method is proposed. In this method, a pre-trained deep learning network and transfer learning are utilized. In addition to fine-tuning and data augmentation, the transfer learning is applied to AlexNet by replacing the last layer by a Softmax to classify three different lesions (melanoma, common nevus and atypical nevus). The proposed model is trained and tested using the ph2 dataset.